Use identity-hashset instead of hashset for approx_count_distinct
