+++ src/daft-core/src/utils/identity_hash_set.rs
@@ -1,0 +1,23 @@
+use std::hash::{BuildHasherDefault, Hasher};
+
+pub type IdentityBuildHasher = BuildHasherDefault<IdentityHasher>;
+
+#[derive(Default)]
+pub struct IdentityHasher {
+    hash: u64,
+}
+
+impl Hasher for IdentityHasher {
+    fn finish(&self) -> u64 {
+        self.hash
+    }
+
+    fn write(&mut self, _bytes: &[u8]) {
+        unreachable!("IdentityHasher should be used by u64")
+    }
+
+    #[inline]
+    fn write_u64(&mut self, i: u64) {
+        self.hash = i;
+    }
+}
