+++ src/daft-table/src/probe_table/mod.rs
@@ -3,13 +3,11 @@ use std::collections::{hash_map::RawEntryMut, HashMap};
 use common_error::DaftResult;
 
 use daft_core::utils::dyn_compare::{build_dyn_multi_array_compare, MultiDynArrayComparator};
+use daft_core::utils::identity_hash_set::IdentityBuildHasher;
 
 use daft_core::{array::ops::as_arrow::AsArrow, schema::SchemaRef};
 
-use crate::{
-    ops::hash::{IdentityBuildHasher, IndexHash},
-    Table,
-};
+use crate::{ops::hash::IndexHash, Table};
 
 struct ArrowTableEntry(Vec<Box<dyn arrow2::array::Array>>);
